dont check for tracking uri when running mlflow offline (#146)
* no longer checking tracking uri when running mlflow offline

old behaviour when running offline with the default tracking uri (???) would cause a crash. My understanding is there is no reason to check tracking uri when running offline because this is provided later when the run is synced to mlflow. One use case for this would be for external vendors benchmarking anemoi. They could use the mlflow logging features to analyse GPU utilization etc. without requiring a tracking uri.

* Update src/anemoi/training/diagnostics/logger.py

Co-authored-by: Harrison Cook <Harrison.cook@ecmwf.int>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: Harrison Cook <Harrison.cook@ecmwf.int>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>