feat(models,training)!: temporal interpolation (#153)
# Adds temporal interpolation functionality to anemoi

A forecaster with longer time steps might yield better predictions going days out than a 1 hour forecaster, as it has to make fewer auto-regressive steps. To produce the hourly predictions still, we can use the information available from a longer time-stepping forecaster, e.g. hours 12 and 18 as input to predict hours 13-17. These predictions can be made individually, assisted by some information about the target time as input.

Preliminary results indicate that this approach works well when combining a 6 hour forecaster and hourly interpolator during inference. Additionally, when hourly analysis is not available, there does not seem to be a major loss in training the interpolator as an emulator on nwp model runs, as long as it is combined with a forecasting model trained on analysis during inference.

---------------------------------------------------------------------------------------------------
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Harrison Cook <harrison.cook@ecmwf.int>