feat!: generalise activation function (#163)
## Description

We want to make the activation function more flexible to allow adding
those that are not in pytorch. In this situation the GLU derivated
(SwiGLU, ReGLU, GEGLU) would be interesting to test.

There is a pytorch implementation of the GLU function, but it splits the
input in 2 parts and make the output half the size of the input. We then
have issue with input size of the next layer. The suggested solution for
that is to make the linear transformation on the complete initial
matrix, but other options are possible and could replace that one.

```py
W = nn.Linear(dim, dim)
V = nn.Linear(dim, dim)

GLU = Sigmoid(W(input)) * V(input)
```

Equation: GLU(x, W, V, b, c) = σ(xW + b) ⊗ (xV + c)

---------

Co-authored-by: Jesper Dramsch <jesper@dramsch.net>
Co-authored-by: Jesper Dramsch <jesper.dramsch@ecmwf.int>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>