test: add test for lam training with existing graph (#336)
## Description
This adds an integration tests in training to run a test cycle with a
lam grid and an existing graph.

## What problem does this change solve?
We test if training works when loading a graph from disk. This is a
feature used often for lam/stretched grid use cases.

## What issue or task does this change relate to?
Integration tests for training

##  Additional notes ##
We are looking into a better testing configuration for the graphs used
in lam/stretched grids separately. This is just to test the
functionality of loading a graph during training.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***