revert(model,training)!: Remove multimapper (#268)
## Description

**Problem:**
For the multimapping (for example mwd to sin(mwd) and cos(mwd)), we need
to create space for the new columns where the variables are mapped to
Suboptimal solution on main:
- create a new batch
- this reallocates also the columns that are not touched
- we need memory for two batches! this does not necessarily cause
problems at o96 and n320 but leads to problems at 9km scale since the
batch takes up ~8GB

**Solution:**
Anemoi-inference now gives the flexibility to allow multimappings
compared to a few months ago, when this was more complex. During dataset
building the multimapping will be done by anemoi-transform
https://github.com/ecmwf/anemoi-transform/pull/31.

Issue https://github.com/ecmwf/aifs-dev/issues/25 shows nicely how the
model needs less memory since the batch is not copied by the
preprocessing in validation anymore.

**Breaking Change**
If the multimapper is used, this needs to be moved to the dataset
creation as the multimapping in the model will not be supported anymore
because it leads to unnecessary memory usage.
**Monomapper is still supported!** (introduced in PR
https://github.com/ecmwf/anemoi-models/pull/88) The naming `monomapper`
will disappear, and we will refer to it as `remapper` only without
distinguishing between mono and multimapping anymore.

**Config file changes**
`data.remapped` needs to be removed from all config files, as this
remapping functionality is no longer supported by the data indices.

## Type of Change

-   [ ] Bug fix (non-breaking change which fixes an issue)
-   [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
-   [x] Documentation update

## Issue Number

Closes https://github.com/ecmwf/aifs-dev/issues/25

## Code Compatibility

-   [x] I have performed a self-review of my code

### Code Performance and Testing

- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I ran the [complete Pytest
test](https://anemoi.readthedocs.io/projects/training/en/latest/dev/testing.html)
suite locally, and they pass
-   [x] I have tested the changes on a single GPU
-   [x] I have tested the changes on multiple GPUs / multi-node setups
- [x] I have run the Benchmark Profiler against the old version of the
code
- [x] If the new feature introduces modifications at the config level, I
have made sure to update Pydantic Schemas and default configs
accordingly

### Dependencies

- [x] I have ensured that the code is still pip-installable after the
changes and runs
- [ ] I have tested that new dependencies themselves are
pip-installable.
- [x] I have not introduced new dependencies in the inference portion of
the pipeline

### Documentation

-   [x] My code follows the style guidelines of this project
- [x] I have updated the documentation and docstrings to reflect the
changes
- [x] I have added comments to my code, particularly in
hard-to-understand areas

## Additional Notes

Should not be merged before changes are in main branch of
anemoi-inference:
https://github.com/ecmwf/anemoi-inference/tree/fix/waves

Possible reviewers: @japols 


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--268.org.readthedocs.build/en/268/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--268.org.readthedocs.build/en/268/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--268.org.readthedocs.build/en/268/

<!-- readthedocs-preview anemoi-models end -->