ci: Synchronised releases (#353)
We currently release each core package separately. At the repo level the
tests ensure compatibility of the state of main, but in the case of
breaking changes a person has to decide which core packages to release
together so that pypi releases remain compatible.

It has happened a few times now where we released models without
realising it broke the pypi version of training.

This PR turns on synchronised releases: instead of 3 PRs, there will now
be 1 PR to trigger a release of all packages that have been changed.
Each core package will still have its own independent version number,
and it should(tm) not create a release for a package if it has not
changed since the last release.

The result will be that every latest pypi release mirrors the state of
main at the time of release, thus ensuring compatibility between all
core packages.