fix: Plotting error when precip related params are not diagnostic  (#369)
## Description
<!-- What issue or task does this change relate to? -->
#368 

## What problem does this change solve?
#368 

## What issue or task does this change relate to?
#368

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***