refactor: use downloading fixture from utils (#372)
## Description
This PR changes the use of downloading utilities for testing according
to the changes made in https://github.com/ecmwf/anemoi-utils/pull/166.
It furthermore updates the fixtures for integration tests so that we
reuse config fixtures for config validation and training cycles.

## What issue or task does this change relate to?
Related to https://github.com/ecmwf/anemoi-utils/issues/147

##  Additional notes ##
The PR in anemoi-utils is merged and the dep bumped accordingly.
The integration tests on this branch pass 
on the hpc:
https://github.com/ecmwf/anemoi-core/actions/runs/15878850545
and on github runners:
https://github.com/ecmwf/anemoi-core/actions/runs/15878838674