remove --n-migrations and --target
sync should be the only entry point. target and n-migration is not
useful if we can migrate to the current migration state (the migration
available for the current version).

As such, migrating to a specific version is as simple as updating your
anemoi-models to a given version, and calling `anemoi-models migration sync`.