feat: Logging to AzureML with ML Flow (#469)
## Description

I have been logging metrics to Azure ML, which requires Azure's special
blend of mlflow. This PR adds those features, which I think might be
useful to a broader audience of people who either run on Azure, Azure
ML, or simply log and view metrics on AML. I'm currently running on an
on-prem HPC machine but logging to AML with this code, however I think
it will be necessary for those actually running on AML as well.

cc @mchantry 

## What problem does this change solve?

Using mlflow to log to Azure ML requires some azure-specific code and
packages.

## What issue or task does this change relate to?

No open issue, please let me know if you'd like me to open one up to
describe the situation, I'd be happy to do so.

##  Additional notes ##

I got this working *just enough* to my liking, but there are still some
issues unfortunately.

* tests: I'm not sure how to make tests with this, since it would
require logging to AML... If anyone has any ideas I'm all ears
* TokenAuth: is not necessary with AML, see necessary helper PR at
ecmwf/anemoi-utils#200
* terminal logs: with regular mlflow, terminal logs get updated by
simply overwriting the terminal_log.txt. Overwriting files is not
allowed with azure-mlflow. Moreover this appears to happen on each
process (?) that's running, so many times over and over. I put in some
try/except logic to try writing these files, but just pass whenever a
failure happens. This means the terminal_log.txt stops updating pretty
early on in the training process. I'm not really sure what to recommend
here.
* diagnostic plots: these also seem to get written by each process
(worker?). However, since each unique plot has a unique name, putting
the try/except logic here doesn't really matter - each process tries to
log the plot and if it succeeds, that means it was the first one and the
plot gets stored. If the process fails, then it just moves on passively
since a different one beat it to the job. no problem but maybe not the
cleanest solution.
* model checkpoints: I simply could not get model checkpoints to log to
AML, and I don't know what is going on there. For my purposes, I don't
really care though, since they get stored locally on the machine where I
will use the checkpoints anyway. Also, @mariahpope at EPIC/NOAA has been
kicking off some anemoi training runs on AML, and noticed the
checkpoints are stored. So, this may not be a problem.
* hyperparameter logging: in the second half of the method
`log_hyperparams_in_mlflow`, the logger tries to log "expanded_params"
after truncating them. I have no idea why but I could not get this to
work for AML, despite lots of attempts with ChatGPT and Microsoft
Copilot. This is the reason for overriding this method in the Azure
subclass.
* I think there are a lot of different ways to establish a connection to
an AML workspace, I did it here with the following code block:

```python
        sp_auth = ServicePrincipalAuthentication(
            tenant_id=os.environ["AZURE_TENANT_ID"],
            service_principal_id=os.environ["AZURE_CLIENT_ID"],
            service_principal_password=os.environ["AZURE_CLIENT_SECRET"],
        )
        ws = Workspace(
            subscription_id=os.environ["AZURE_SUBSCRIPTION_ID"],
            resource_group=aml_resource_group,
            workspace_name=aml_workspace_name,
            auth=sp_auth,
        )
```

Note there's a check before those lines to make sure the necessary
environment variables are defined for the user. I would be very open to
making this more flexible based on what true AML users think, or we
could let this evolve as more users start to flex this feature.


Let me know if these limitations are acceptable, and I'd be happy to put
some warning flags for the user so it's clear what will and will not be
logged to AML. If anyone wants these to be worked out in the PR, I'd
love to hear some suggestions - I'm pretty new to AML and leaned heavily
on some LLMs to make this work, but "we" all got stuck on these issues
:)


---

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Nathan Simpson <nsimpson@turing.ac.uk>
Co-authored-by: Nathan Simpson <phinate@pm.me>
Co-authored-by: Nathan Simpson <49782545+phinate@users.noreply.github.com>
Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>