feat: activate minmium plotting for integration tests (#669)
## Description
Testing plots it's not part our current testing suite. Since most of
them are implemented via callbacks a simple way would be to activate a
minimal plotting set up for the integration test to ensure we don't
break them in the future.

This PR proposes those changes:
- Spectrum plot it's quite slow so for now it's deactivated for the SG
model
- Since the test dataset for LAM and SG is a small one we just test
plotting tp
- The plots for the interpolator are deactivated - while the callbacks
run the output produced is not scientifically correct. More work will be
done in the future to address this (@mc4117 - notified about this in
slack channel)
- In the hierarchical model - current plotting of edge features is not
support - this plot is for now deactivated. (Discussed with @icedoom888
)

cc: @VeraChristina - having those tests would also be useful in the
context of multiple datasets to ensure there is no feature regression
for plotting.


## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->

## What issue or task does this change relate to?
<!-- link to Issue Number -->

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: Mariana Clare <mariana.clare@ecmwf.int>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>