fix: Target docs (#704)
Replace ``_target`` with ``_target_``.
Fix YAML indentation.

## Description
<!-- What issue or task does this change relate to? -->
Typo(s) in the docs.

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->
Bugfix

## What issue or task does this change relate to?
<!-- link to Issue Number -->
N/A

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->
I think that this reference to `_target` is a typo, which should say
`_target_`.
I believe that some YAML blocks are not valid (as tested with
`yaml.safe_load(the_text)`) and that YAML objects can either be lists or
dicts but not a mixture of both.
Not valid (raises ParserError):

```yaml
a_key:
  - a list item
  another_key: a value
```

Valid:

```yaml
a_key:
- a_list_of_dicts_key_1: val
  a_list_of_dicts_key_2: val
```

[rstcheck](https://github.com/rstcheck/rstcheck/tree/main) catches these
(and a few others). I'd be happy to add it to the CI system if that's
something you'd be interested in.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--704.org.readthedocs.build/en/704/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--704.org.readthedocs.build/en/704/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--704.org.readthedocs.build/en/704/

<!-- readthedocs-preview anemoi-models end -->

---------

Co-authored-by: Matthew Chantry <matthew.chantry@ecmwf.int>