fix(docs,schemas): update hardware reference in docstrings, documentation and schemas (#737)
## Description
When updating the configs for the system section, we forgot to update
some details regarding the docs. This PR fixes that and some other
things that Mario reported in
https://github.com/ecmwf/anemoi-core/issues/703

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->

## What issue or task does this change relate to?
[<!-- link to Issue Number -->
](https://github.com/ecmwf/anemoi-core/issues/703)

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: Mario Santa Cruz <mariosanta_cruz@hotmail.com>