fix: progress bar with lightning==2.6.0 (#739)
With the new version of lightning (2.6.0) the default changes to
RichProgressBar if the rich package is available
(https://github.com/Lightning-AI/pytorch-lightning/releases/tag/2.6.0).
On HPCs with Slurm, where the output is written to a file, this does not
work.

This introduces a progress bar callback, which allows you to use the
different lightning progress bars or your custom progress bar. The
progress bar class can be set in the callbacks and will be instantiated,
e.g.
```
progress_bar:
  _target_: pytorch_lightning.callbacks.TQDMProgressBar
  refresh_rate: 1
```
If no `progress_bar` is set, or the instantiation fails, a default
`pytorch_lightning.callbacks.TQDMProgressBar` will be used. This also
ensures backwards compatibility with the configs.

Tests with different lighting versions (2.5.2, 2.5.5, 2.6.0) and both
training interactively and with a slurm job worked well with the
`pytorch_lightning.callbacks.TQDMProgressBar`.

- Interactive debug run with RichColorbar:
https://mlflow.ecmwf.int/#/experiments/395/runs/266ea7709a2a476e9337a60079f151b8/artifacts
- Interactive debug run with TQDMProgressBar:
https://mlflow.ecmwf.int/#/experiments/395/runs/b417e46603974cbc80823938434b1103/artifacts
- Slurm debug run with TQDMProgressBar:
https://mlflow.ecmwf.int/#/experiments/413/runs/8d2ecd0b1c68429b881fadc5db1bf51d/artifacts
- Slurm debug run with RichProgressBar:
https://mlflow.ecmwf.int/#/experiments/413/runs/d37e47b2ed814aacb607e6ccc7a5ec37/artifacts

**Note:** The `pytorch_lightning.callbacks.RichProgressBar` does not
work on slurm jobs. The progress bar only shows after the end of
training.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚:
https://anemoi-training--739.org.readthedocs.build/en/739/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚:
https://anemoi-graphs--739.org.readthedocs.build/en/739/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚:
https://anemoi-models--739.org.readthedocs.build/en/739/

<!-- readthedocs-preview anemoi-models end -->

---------
Co-authored-by: anaprietonem <ana.prietonemesio@ecmwf.int>