feat(models): triton gt more shapes (#752)
## Description
Triton GT can now run where number of heads and/or number of channels !=
a power of 2. e.g. 1526 channels or 12 heads

This is achieved by padding the matrix offsets to reach the power of 2
size, and then masking out the extra values before loading or storing.
This makes the code a bit uglier to follow, but i added some comments.
Also I had to allocate some temp arrays with shape (H_pad, C_pad), so
the memory usage isnt ideal for non-power-of-2 inputs

i split the pytests into forward and backward reference tests, and used
a different function to test correctness which tells you if you're 1%
off or 99% off the reference (which can be very motivating when
debugging :D)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>